add filter, solve delay issue when adding or toggling
