--- samples/subapp2-react-query-todoapp/src/routes.ts
@@ -1,8 +1,0 @@
-export const favicon = "static/favicon.png";
-
-export default {
-  "/": {
-    pageTitle: "Welcome to Electrode",
-    subApps: ["./home", "./demo1", "./demo2"]
-  }
-};
