+++ samples/subapp2-react-query-todoapp/src/todo/index.tsx
@@ -1,9 +1,6 @@
 import { React, ReactSubApp, createDynamicComponent } from "@xarc/react";
-import { reduxFeature } from "@xarc/react-redux";
-import { VISIBILITY_FILTERS } from "./constant";
 
-export { reduxReducers } from "./redux/reducer";
-import { reactQueryFeature, useQuery } from "@xarc/react-query";
+import { reactQueryFeature } from "@xarc/react-query";
 
 export const TodoApp = createDynamicComponent({
   name: "todoapp",
